+++ Configuration Management/ConnectionSettingsWindow.xaml.cs
@@ -1,4 +1,6 @@
+#if WINDOWS
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Windows;
 using System.Windows.Controls;
@@ -332,3 +334,4 @@ namespace Configuration_Management
         }
     }
 }
+#endif