+++ Configuration Management/LaunchParametersWindow.xaml.cs
@@ -1,3 +1,4 @@
+#if WINDOWS
 using System.Windows;
 using System.Windows.Input;
 using Configuration_Management.Localization;
@@ -150,3 +151,4 @@ namespace Configuration_Management
         }
     }
 }
+#endif