+++ Configuration Management/LinkInputWindow.xaml.cs
@@ -1,3 +1,4 @@
+#if WINDOWS
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
@@ -49,3 +50,4 @@ namespace Configuration_Management
         }
     }
 }
+#endif