+++ Configuration Management/NameInputWindow.xaml.cs
@@ -1,3 +1,4 @@
+#if WINDOWS
 using System.Windows;
 using System.Windows.Input;
 
@@ -51,3 +52,4 @@ namespace Configuration_Management
         }
     }
 }
+#endif