+++ Configuration Management/TagInputWindow.xaml.cs
@@ -1,3 +1,4 @@
+#if WINDOWS
 using System.Windows;
 using System.Windows.Input;
 
@@ -42,3 +43,4 @@ namespace Configuration_Management
         }
     }
 }
+#endif